update mac info and imager

+++ cloudmesh/burn/Imager.py
@@ -1,6 +1,11 @@
 import os
 from cloudmesh.common.Shell import Shell
 
+from cloudmesh.burn.image import Image
+from cloudmesh.burn.util import os_is_linux
+from cloudmesh.burn.util import os_is_pi
+from cloudmesh.common.console import Console
+
 class Imager:
 
 
@@ -9,26 +14,29 @@ class Imager:
         return r is not None
     
     @staticmethod
-    def install():
-        # TODO: if installed skip
-        os.system("sudo apt install -y rpi-imager")
+    def install(force=False):
+        if not installed or force:
+            if os_is_linux() or os_is_pi():
+                os.system("sudo apt uninstall -y rpi-imager")
+            else:
+                Console.warning("Instalation is not supported")
+    @staticmethod
+    def download(tag=["latest-lite"]):
+        pass
 
     @staticmethod
-    def install(force=false):
-        if not installed or force:
-            os.system("sudo apt uninstall -y rpi-imager")
-
     def format(file=None):
 
         if file is not None:
             if not os.path.exists(file):
-                raise ValueError("image file does not exist")
+                raise ValueError(f"image file {file} does not exist")
 
-            if not file.endswith(".img")
-               raise ValueError("file {file} does not end with .img")
+            if not file.endswith(".img"):
+               raise ValueError(f"file {file} does not end with .img")
             
         elif file is None:
             file = ""
             
         Imager.install()
-        os.system("sudo rpi-imager {file}")
+
+        os.system(f"sudo rpi-imager {file}")
